refactor: 게임 참가자가 직접 수익이 얼마인지 알려줄 수 있도록 수정

+++ src/main/kotlin/blackjack/domain/model/progress/ProfitStatistics.kt
@@ -7,7 +7,7 @@ class ProfitStatistics(
     private val betHistory: BetHistory,
     private val winLossStatistics: WinLossStatistics,
 ) {
-    val playerProfits: Map<Player, Float>
+    val playerProfits: Map<Player, Double>
         get() {
             val playerWinLoseInfo = winLossStatistics.playerWinLoseInfo
             return betHistory.playerBets.entries.associate { (player, betAmount) ->
@@ -18,22 +18,22 @@ class ProfitStatistics(
                 player to calculateProfit(betAmount, profitRate)
             }
         }
-    val dealerProfits: Float
-        get() = -playerProfits.values.sum() + 0.0f // -0 방지
+    val dealerProfits: Double
+        get() = -playerProfits.values.sum() + 0.0 // -0 방지
 
     private fun calculateProfit(
         betAmount: Int,
-        profitRate: Float,
-    ): Float = betAmount * profitRate
+        profitRate: Double,
+    ): Double = betAmount * profitRate
 
     private fun calculateProfitRate(
         winLoss: WinLoss,
         cardStatus: CardStatus,
-    ): Float =
+    ): Double =
         when (winLoss) {
-            WinLoss.WIN -> if (cardStatus == CardStatus.BLACKJACK) 1.5f else 1f
-            WinLoss.LOSE -> -1f
-            WinLoss.DRAW -> 0f
+            WinLoss.WIN -> if (cardStatus == CardStatus.BLACKJACK) 1.5 else 1.0
+            WinLoss.LOSE -> -1.0
+            WinLoss.DRAW -> 0.0
         }
 
     companion object {
